Move all dependency from moonstreamdb to networks. Will required upper version of moonstreamdb.
